Fixes project loading error related to gates, which do not reference/use a gate template.
This is how the error happened:
1. Open existing project.
2. Remove arbitrary gate template from the template library.
3. Select to KEEP gates, which have been using the removed template.
   This will set these gates' template type identifiers to 0.
4. Save project.
5. Reload project.
6. => degate refuses to load the project.

The error occurred, because degate did not properly handle gates, which
do not reference/use any template. In such a case, degate unsuccessfully
attempted to retrieve the gate template with id 0; of course, such a template
does not exist and an InvalidObjectIDException is thrown.

This commit fixes the issue by checking for gate_type_id == 0 before
calling gate_library->get_template.